+++ docs/CHANGELOG.rst
@@ -26,6 +26,8 @@ Changed
 Fixed
 -----
 - Fix multi-file download stopped after downolading html file
+- Modify ``clear_cache`` method of ``BaseTables`` to clear only the
+  cache of the specified collection
 
 
 ===================